# src/main/resources/apidoc.json
{"openapi":"3.0.1","info":{"title":"Deduction Management API products","description":"This documentation exposes contracts to deduction management APIs .","termsOfService":"https://www.prgx.com/terms","contact":{"name":"prgx","url":"https://www.prgx.com","email":"support@prgx.com"},"version":"1.0"},"servers":[{"url":"http://learn.supplierconnect.prgx.com:8080","description":"Server URL in Learn environment"},{"url":"https://supplierconnect.prgx.com","description":"Server URL in Production environment"}],"paths":{"/api/promos":{"get":{"tags":["Promo Allowances"],"summary":"Retrieve all Promo Allowances","operationId":"getAllPromo","parameters":[{"name":"startDate","in":"query","required":true,"schema":{"type":"string"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string"}},{"name":"supplierName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Promo"}}}},"500":{"description":"Internal Server Error"},"204":{"description":"There are no promo allowances to report"}}}},"/api/promos/{promoID}":{"get":{"tags":["Promo Allowances"],"summary":"Retrieve Promo Line Details by Promo ID","description":"Get a Promo Line object by specifying its id. The response is PromoLine object.","operationId":"getPromoById","parameters":[{"name":"promoID","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error"},"404":{"description":"Not Found"},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromoLine"}}}}}}},"/api/payments":{"get":{"tags":["Payment"],"summary":"Retrieve all payments","operationId":"getAllPayments","parameters":[{"name":"startDate","in":"query","required":true,"schema":{"type":"string"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string"}},{"name":"supplierName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error"},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Payment"}}}},"204":{"description":"There are no payments"}}}},"/api/invoices":{"get":{"tags":["Invoice"],"summary":"Retrieve all Invoices","operationId":"getAllInvoices","parameters":[{"name":"startDate","in":"query","required":true,"schema":{"type":"string"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string"}},{"name":"supplierName","in":"query","required":true,"schema":{"type":"string"}},{"name":"invoiceStatus","in":"query","required":false,"schema":{"type":"string"}},{"name":"storeNumber","in":"query","required":false,"schema":{"type":"string"}},{"name":"invoiceCategory","in":"query","required":false,"schema":{"type":"string"}},{"name":"associtaedDeductions","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error"},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Invoice"}}}},"204":{"description":"There are no Invoice"}}}},"/api/invoices/{invoiceNumber}":{"get":{"tags":["Invoice"],"summary":"Retrieve Invoice Details by Invoice Number","description":"Get a Invoice object by specifying its id. The response is InvoiceHeader object with InvoiceLines.","operationId":"getInvoiceById","parameters":[{"name":"invoiceNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error"},"404":{"description":"Not Found"},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvoiceDetail"}}}}}}},"/api/coupons":{"get":{"tags":["Coupon Deductions"],"summary":"Retrieve all Coupon Deductions","operationId":"getAllCoupons","parameters":[{"name":"startDate","in":"query","required":true,"schema":{"type":"string"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string"}},{"name":"supplierName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error"},"204":{"description":"There are no coupon deductions"},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CouponDeductionLine"}}}}}}},"/api/coupons/{couponDeductionNo}":{"get":{"tags":["Coupon Deductions"],"summary":"Retrieve Coupon Deduction Line Details by ID","description":"Get a Deduction Line object by specifying its id. The response is CouponDeductionLine object.","operationId":"getCouponById","parameters":[{"name":"couponDeductionNo","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error"},"404":{"description":"Not Found"},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CouponDeduction"}}}}}}},"/api/claims":{"get":{"tags":["Claims"],"summary":"Retrieve all Claims","operationId":"getAllClaims","parameters":[{"name":"startDate","in":"query","required":true,"schema":{"type":"string"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string"}},{"name":"supplierName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"There are no disputes to report"},"500":{"description":"Internal Server Error"},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Claim"}}}}}}}},"components":{"schemas":{"Promo":{"type":"object","properties":{"invoiceNumber":{"type":"string","description":"Invoice Number ","example":"AK-6055237"},"dealNumber":{"type":"string","description":"Deal Number ","example":"470718"},"contractNumber":{"type":"string","description":"Contract Number ","example":"X70718"},"offerNumber":{"type":"string","description":"Offer Number ","example":"X70718"},"promoID":{"type":"integer","description":"Promo ID ","format":"int64","example":57177802},"promoDivision":{"type":"string","description":"Promo Division ","example":"0016-Columbus Operating Division"},"promoInvoiceAmount":{"type":"integer","description":"Promo Invoice Amount ","format":"int64"},"invoiceDate":{"type":"string","description":"Invoice Date ","format":"date-time"},"supplier":{"$ref":"#/components/schemas/Supplier"},"noOfPayments":{"type":"integer","description":"Number of Payments ","format":"int64","example":10},"noOfCliams":{"type":"integer","description":"Number of Claims ","format":"int64","example":10},"paidAmount":{"type":"integer","description":"Total Paid Amount ","format":"int64"},"currency":{"type":"string","description":"Currency","example":"USD"},"poNumber":{"type":"string","description":"PO Number"},"storeNumber":{"type":"string","description":"Store Number"}},"description":"Promo allowances associated to Invoice "},"Supplier":{"type":"object","properties":{"supplierERPID":{"type":"integer","format":"int64"},"supplierName":{"type":"string"},"paymentRefNo":{"type":"integer","format":"int64"}},"description":"Supplier"},"Invoice":{"type":"object","properties":{"Invoice Number":{"type":"string","example":"092-A447148-410"},"supplier":{"$ref":"#/components/schemas/Supplier"},"invoiceDate":{"type":"string"},"division":{"type":"string"},"storeNumber":{"type":"string"},"netInvoiceAmount":{"type":"number","format":"double"},"totalDeductionAmount":{"type":"integer","format":"int64"},"grossInvoiceAmount":{"type":"number","format":"double"},"totalPaidAmount":{"type":"number","format":"double"},"currency":{"type":"string"},"invoiceStatus":{"type":"string"},"invoiceLocation":{"type":"string"}},"description":"Invoice Details"},"PromoLine":{"type":"object","properties":{"promoHeader":{"$ref":"#/components/schemas/Promo"},"supplier":{"$ref":"#/components/schemas/Supplier"},"invoice":{"$ref":"#/components/schemas/Invoice"},"fiscalYear":{"type":"string","description":"Fiscal Year","example":"2023"},"fiscalPeriod":{"type":"string","description":"Fiscal Period","example":"8"},"weekEndingDate":{"type":"string","description":"Week Ending Date ","format":"date-time"},"promoInvoiceAmount":{"type":"integer","description":"Promo Invoice Amount ","format":"int64"},"accountInvoiceDate":{"type":"string","description":"Accounting Invoice Date ","format":"date-time"},"contractDueAmount":{"type":"integer","description":"Contract due amount","format":"int64","example":10},"dealNumber":{"type":"integer","description":"Deal Number ","format":"int64","example":10},"contractNumber":{"type":"integer","description":"Contract Number ","format":"int64","example":10},"linePromoDivision":{"type":"integer","description":"Line Promo Division ","format":"int64"},"promoDivision":{"type":"integer","description":"Promo Division ","format":"int64"},"divisionNumber":{"type":"integer","description":"Division Number ","format":"int64"},"divisionName":{"type":"integer","description":"Division Name ","format":"int64"},"contractStartDate":{"type":"string","description":"Contract Start Date ","format":"date-time"},"contractEndDate":{"type":"string","description":"Contract end Date ","format":"date-time"},"couponNumber":{"type":"string","description":"Coupon Number"},"couponPercentage":{"type":"string","description":"Coupon Billing Percentage"},"offerCode":{"type":"string","description":"Offer Code"},"eventID":{"type":"string","description":"Event ID"},"mfgOffer":{"type":"string","description":"Manufacture Offer"},"mfgCode":{"type":"string","description":"Manufacture Code"},"couponDesc":{"type":"string","description":"Coupon Description"},"userComments":{"type":"string","description":"User Comments"},"contractType":{"type":"string","description":"Contract Type"},"categoryManager":{"type":"string","description":"Category Manager "},"lineNo":{"type":"string","description":"Line Number "},"upc":{"type":"string","description":"upc"},"location":{"type":"string","description":"Location"},"itemDescription":{"type":"string","description":"Item Description"}}},"Payment":{"type":"object","properties":{"paymentRefNumber":{"type":"string","description":"Payment Reference Number","example":"100956996"},"settlementNumber":{"type":"string","description":"Settlement number","example":"3709929"},"remittanceMethod":{"type":"string","description":"Remittance method like ERP/ORA_AP/AR Netting","example":"EFT"},"invoiceNumber":{"type":"string","description":"Invoice number","example":"092-A447148-410"},"invoiceDate":{"type":"string","description":"Invoice Date","format":"date-time"},"paymentCurrency":{"type":"string","description":"Payment Currency","example":"USD"},"invoiceCurrency":{"type":"string","description":"Invoice Currency","example":"USD"},"grossInvoiceAmount":{"type":"number","description":"Gross Invoice Amount","format":"double","example":100},"totalDeductionAmount":{"type":"number","description":"Total Deduction Amount","format":"double","example":100.12},"totalDiscountAmount":{"type":"number","description":"Total Discount Amount","format":"double","example":200.9},"netInvoiceAmount":{"type":"number","description":"Net Invoice Amount","format":"double","example":200.9},"paidAmount":{"type":"number","description":"Paid Amount","format":"double","example":200.9},"checkAmount":{"type":"number","description":"Check Amount","format":"double","example":7724.39},"CheckStatus":{"type":"number","description":"Check Status","format":"double"},"paymentRefDate":{"type":"number","description":"paymentRefDate","format":"double"},"numberOfClaims":{"type":"number","description":"numberOfClaims","format":"double","example":1},"PONumber":{"type":"number","description":"PO Number","format":"double"},"payeeNumber":{"type":"number","description":"Payee Number","format":"double"},"division":{"type":"number","description":"Division ","format":"double"},"stroreNumber":{"type":"number","description":"Strore Number ","format":"double"},"legacyStoreNumber":{"type":"number","description":"Legacy Store Number ","format":"double"},"supplierERPID":{"type":"number","description":"Supplier ERP ID ","format":"double"},"supplierName":{"type":"number","description":"Supplier Name ","format":"double"}},"description":"Payments associated to Invoice"},"Claim":{"type":"object","properties":{"claimID":{"type":"integer","description":"Claim ID","format":"int64","example":15375},"claimType":{"type":"string","description":"Claim Type","example":"LC-List Cost"},"claimCreatedBy":{"type":"integer","description":"Claim created by","format":"int64"},"claimAmount":{"type":"integer","description":"Claim Amount","format":"int64","example":1000},"currency":{"type":"string","description":"Currency","example":"USD"},"claimResolution":{"type":"string","description":"Claim Resolution ","format":"date-time"},"claimStatus":{"type":"string","description":"Claim Status ","format":"date-time"},"paymentRefNo":{"type":"string","description":"Payment Reference Number ","format":"date-time"},"paidAmt":{"type":"string","description":"Paid Amount ","format":"date-time"},"invoiceNumber":{"type":"string","description":"Invoice Number ","example":"AK-6055237"},"invoiceDate":{"type":"string","description":"Invoice Date ","format":"date-time"},"division":{"type":"string","description":"Division ","format":"date-time"},"totalAmount":{"type":"integer","description":"Total invoice amount","format":"int64"},"supplier":{"$ref":"#/components/schemas/Supplier"}}},"DeductionLine":{"type":"object","properties":{"supplier":{"$ref":"#/components/schemas/Supplier"},"paymentRefNo":{"type":"integer","description":"Payment reference number","format":"int64"},"invoice Number":{"type":"integer","format":"int64"},"totalInvoiceAmount":{"type":"integer","format":"int64"},"deductionAmount":{"type":"integer","format":"int64"},"discountAmountt":{"type":"number","format":"double"},"paidAmount":{"type":"number","format":"double"},"currency":{"type":"string"},"paymentRefDate":{"type":"string"},"netInvoiceAmount":{"type":"number","format":"double"},"poNumber":{"type":"integer","format":"int64"},"division":{"type":"string"},"invoiceDate":{"type":"string"},"invoiceLocation":{"type":"string"},"upc":{"type":"integer","format":"int64"},"itemDescription":{"type":"string"},"adjQty":{"type":"string"},"unitPrice":{"type":"number","format":"double"},"adjAmt":{"type":"integer","format":"int64"},"code":{"type":"string"},"adjustmentReason":{"type":"string"},"poCreationDate":{"type":"string"},"originalInvoiceNumber":{"type":"string"},"sourcePoNumber":{"type":"string"},"storeName":{"type":"string"}}},"InvoiceDetail":{"type":"object","properties":{"invoiceSummary":{"$ref":"#/components/schemas/Invoice"},"deductions ":{"type":"array","items":{"$ref":"#/components/schemas/DeductionLine"}},"payments":{"type":"array","description":"Payments associated to Invoice","items":{"$ref":"#/components/schemas/Payment"}},"promos":{"type":"array","description":"Promo allowances associated to Invoice ","items":{"$ref":"#/components/schemas/Promo"}},"claims":{"type":"array","description":"Claims associated to Invoice ","items":{"$ref":"#/components/schemas/Claim"}}}},"CouponDeduction":{"type":"object","properties":{"couponDeductionNumber":{"type":"string","description":"Coupon Deduction Number "},"couponIvoiceCnt":{"type":"integer","description":"No of coupon invoices","format":"int64","example":4},"manufacturerNo":{"type":"integer","description":"Coupon manufacturer Number","format":"int64","example":123},"manufacturerName":{"type":"integer","description":"Coupon manufacturer name","format":"int64"},"supplier":{"type":"integer","description":"Supplier Details","format":"int64"},"reportDate":{"type":"string","description":"Report Date","format":"date-time"},"currency":{"type":"string","description":"Currency","format":"date-time"},"deductionAmount":{"type":"string","description":"Deduction Amount","format":"date-time"}}},"CouponDeductionLine":{"type":"object","properties":{"couponDeduction":{"$ref":"#/components/schemas/CouponDeduction"},"supplier":{"$ref":"#/components/schemas/Supplier"},"invoice":{"$ref":"#/components/schemas/Invoice"},"adjustmentCode":{"type":"string","description":"Adjustment Code"},"adjustmentDescription":{"type":"string","description":"Adjustment Description"}}}}}}